Graph Algorithms

+++ go/interfaces/comparable.go
@@ -1,6 +1,6 @@
 package interfaces
 
 type Comparable interface {
-  Compare(other Comparable) int8
-  Value() interface{}
+	Compare(other Comparable) int
+	Value() interface{}
 }
